requirements.txt
zenml==0.75.0
loguru==0.7.3
tqdm==4.67.1
pymongo==4.11.2
beautifulsoup4==4.12.3
numpy==2.2.4
selenium==4.17.2
qdrant-client==1.13.3
sentence-transformers==3.4.1
python-dotenv==1.0.1
google-generativeai==0.8.4
pandas==2.2.3
scikit-learn==1.6.1
unsloth==2025.3.19